feat: base and ref settings should take priority
This change makes the `base` and `ref` settings take priority over
the GitHub Workflow events.  Previously, if the action was run for
a pull request event, these settings were ignored.

The reason for this change is so we can support pull requests of
VERSION manifest files that will compare the manifest's currently
specified tag version with its previously specified tag version as
part of the PR.